requirements.txt
albumentations==1.3.1
glog==0.3.1
matplotlib==3.8.0
numpy==1.24.4
opencv_contrib_python==4.7.0.72
opencv_python==4.11.0.86
opencv_python_headless==4.8.1.78
Pillow==11.2.1
tensorboardX==2.6.4
torch>=2.4.0
torchvision==0.19.0
tqdm==4.66.5
wandb==0.16.1
